// types.ts
export interface User {
  id: string;
  username: string;
  password: string; // In real app, this should be hashed
  fullName: string;
  role: 'admin' | 'student' | 'parent';
  relatedId?: string; // Links to Student ID or Parent ID if applicable
}
// ...
export interface Student {
  id: string;
  classId: string;
  fullName: string;
  dob: string;
  gender: 'Male' | 'Female';
  address: string;
  parentId: string; // Links to Parent
  status: 'Active' | 'Inactive' | 'Suspended';
  
  // App specific fields
  avatar?: string;
  points: number; 
}
// ...
export interface Task {
  id: string;
  classId: string;
  title: string;
  description: string;
  dueDate: string;
  requireReply: boolean;
  createdAt: string;
  attachments?: string[]; // Teacher attachments (URLs)
  // New fields for Folder structure
  grade?: string;
  unit?: string;
}

export interface TaskReply {
  id: string;
  taskId: string;
  studentId: string;
  replyText: string;
  attachmentsJson?: string; // JSON string of links
  submittedAt: string;
  grade?: number;
  feedback?: string;
}